pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "causal-methods-demo"
version = "0.1.0"
description = "Demonstrating causal inference methods for tax software Smart Filing Assistant impact analysis"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Causal Inference Demo"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = [
    # Core data science packages
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "plotly>=5.0.0",
    
    # Machine learning
    "scikit-learn>=1.1.0",
    "xgboost>=1.6.0",
    "lightgbm>=3.3.0",
    
    # Causal inference specific packages
    "econml>=0.14.0",  # Microsoft's causal ML library
    "dowhy>=0.9.0",    # Microsoft's causal inference library
    "causalml>=0.14.0",  # Uber's causal ML library
    "pymc>=5.0.0",     # For Bayesian causal inference
    
    # Statistical packages
    "scipy>=1.9.0",
    "statsmodels>=0.13.0",
    "linearmodels>=4.27",  # For IV and panel data methods
    
    # Jupyter and visualization
    "jupyter>=1.0.0",
    "notebook>=6.4.0",
    "ipywidgets>=7.6.0",
    
    # Data generation and utilities
    "faker>=18.0.0",  # For realistic synthetic data
    "tqdm>=4.64.0",   # Progress bars
    "pyyaml>=6.0",    # For configuration files
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/example/causal-methods-demo"
Repository = "https://github.com/example/causal-methods-demo"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "mypy>=1.17.0",
]
test = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
]

[tool.ruff]
# Set line length to match project preference
line-length = 88
target-version = "py310"

[tool.ruff.lint]
# Enable commonly used rule sets
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "C90", # mccabe
]

# Allow some commonly used long lines in tests and documentation
ignore = [
    "E501",  # Line too long - will be handled by formatter when possible
]

[tool.ruff.lint.per-file-ignores]
# Tests can be a bit more lenient
"tests/*" = [
    "F841",  # Unused variable - common in tests for setup
]

[tool.ruff.format]
# Use double quotes consistently
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 
